backend/requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
sqlalchemy>=2.0.36,<3.0.0
pydantic-settings>=2.6.1,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
passlib>=1.7.4,<2.0.0
langchain>=0.3.9,<1.0.0
python-multipart>=0.0.9,<1.0.0
pypdf>=5.1.0,<6.0.0
python-docx>=1.1.2,<2.0.0
